Changed:
- Patch to fix problem uncovered in Squeak 4.1, Will still work in older Squeaks.

C. David Shaffer to Chris, John
Jun 25/2010
	
Chris and John,

I see that you made the most recent publishes to Cryptography and SSL
(respectively) so I'm sending this to you.  In Squeak4.1 in looks like
WriteStream>>nextPutAll: answers the argument whereas older Squeaks it
answered the stream.  Attached is a fix for a problem that this causes
with the SSL package.  The fix will work with older Squeak's as well.

I'm still having some troubles with SSL in Squeak4.1 but I think it is
my code as all of the simple examples seem to work fine once you make
this patch.